+++ src/test/java/softeer2nd/chess/BoardTest.java
@@ -171,418 +171,4 @@ class BoardTest {
         assertEquals(Queen.createWhiteQueen(), board.getSortedWhitePieces().get(0));
         assertEquals(Rook.createWhiteRook(), board.getSortedWhitePieces().get(1));
     }
-
-    @Test
-    @DisplayName("단순 이동 기물 이동 테스트")
-    void moveNonRecursive() {
-        // given
-        String sample =
-                ".K......" +
-                "........" +
-                "K....KN." +
-                "........" +
-                "........" +
-                "........" +
-                "........" +
-                ".......K";
-
-        String expect =
-                "........\n" +
-                "..K.....\n" +
-                ".....KN.\n" +
-                "........\n" +
-                "K.......\n" +
-                "........\n" +
-                "........\n" +
-                ".......K\n";
-
-        ChessGame.initialize(board, sample);
-
-        // when
-        ChessGame.move(board, new Board.Position("b8"), new Board.Position("b7"));
-        ChessGame.move(board, new Board.Position("b7"), new Board.Position("c7"));
-
-        assertThrows(
-                IllegalArgumentException.class,
-                () -> ChessGame.move(board, new Board.Position("c7"), new Board.Position("a8"))
-        );
-
-        ChessGame.move(board, new Board.Position("a6"), new Board.Position("a5"));
-        ChessGame.move(board, new Board.Position("a5"), new Board.Position("a4"));
-        assertThrows(
-                IllegalArgumentException.class,
-                () -> ChessGame.move(board, new Board.Position("h1"), new Board.Position("h0"))
-        );
-        assertThrows(
-                IllegalArgumentException.class,
-                () -> ChessGame.move(board, new Board.Position("h1"), new Board.Position("i1"))
-        );
-
-        assertThrows(
-                IllegalArgumentException.class,
-                () -> ChessGame.move(board, new Board.Position("f6"), new Board.Position("g6"))
-        );
-
-        // then
-        assertEquals(expect, showBoard(board));
-    }
-
-    @Test
-    @DisplayName("재귀적 이동 기물 이동 테스트")
-    void moveRecursive() {
-        // given
-        String sample =
-                "........" +
-                "........" +
-                "Q....QN." +
-                "........" +
-                "........" +
-                "........" +
-                "........" +
-                ".......Q";
-
-        String expect =
-                "........\n" +
-                "........\n" +
-                "......N.\n" +
-                "........\n" +
-                "....Q...\n" +
-                "Q.......\n" +
-                "........\n" +
-                ".....Q..\n";
-
-        ChessGame.initialize(board, sample);
-
-        // when
-        ChessGame.move(board, new Board.Position("a6"), new Board.Position("a8"));
-        ChessGame.move(board, new Board.Position("a8"), new Board.Position("a6"));
-        ChessGame.move(board, new Board.Position("a6"), new Board.Position("a5"));
-        ChessGame.move(board, new Board.Position("a5"), new Board.Position("a3"));
-
-
-        assertThrows(
-                IllegalArgumentException.class,
-                () -> ChessGame.move(board, new Board.Position("f6"), new Board.Position("g6"))
-        );
-
-        assertThrows(
-                IllegalArgumentException.class,
-                () -> ChessGame.move(board, new Board.Position("h1"), new Board.Position("d4"))
-        );
-        ChessGame.move(board, new Board.Position("h1"), new Board.Position("e4"));
-        ChessGame.move(board, new Board.Position("f6"), new Board.Position("f1"));
-
-
-        // then
-        assertEquals(expect, showBoard(board));
-    }
-
-    @Test
-    @DisplayName("기물을 넘어갈 수 없으며 색깔이 다른 기물 위치로 이동하면 해당 기물이 제거되어야 한다.")
-    void remove() {
-        // given
-        String sample =
-                "........" +
-                "........" +
-                ".q...Q.." +
-                "........" +
-                "........" +
-                "........" +
-                "........" +
-                "........";
-
-        String expect =
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                ".Q......" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE;
-        ChessGame.initialize(board, sample);
-
-        // when
-        assertThrows(
-                IllegalArgumentException.class,
-                () -> ChessGame.move(board, new Board.Position("f6"), new Board.Position("a6"))
-        );
-        ChessGame.move(board, new Board.Position("f6") ,new Board.Position("b6"));
-
-        // then
-        assertEquals(expect, BoardView.showBoard(board));
-    }
-
-    @Test
-    @DisplayName("같은 색 기물은 제거되면 안된다.")
-    void notRemove() {
-        // given
-        String sample =
-                "........" +
-                "........" +
-                ".Q...Q.." +
-                "........" +
-                "........" +
-                "........" +
-                "........" +
-                "........";
-
-        String expect =
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                ".Q...Q.." + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE;
-        ChessGame.initialize(board, sample);
-
-        // when
-        assertThrows(
-                IllegalArgumentException.class,
-                () -> ChessGame.move(board, new Board.Position("f6"), new Board.Position("a6"))
-        );
-
-        // then
-        assertEquals(expect, BoardView.showBoard(board));
-    }
-
-    @Test
-    @DisplayName("룩 기물의 움직임 테스트")
-    void moveRook() {
-        // given
-        String rookSample =
-                "........" +
-                "........" +
-                "....R...." +
-                "........" +
-                "........" +
-                "........" +
-                "........" +
-                "........";
-
-        String expect =
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                ".....R.." + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........"+ NEWLINE;
-        ChessGame.initialize(board, rookSample);
-
-        // when
-        ChessGame.move(board, new Board.Position("e6"), new Board.Position("e8"));
-        Piece result1 = board.findPiece(new Board.Position("e8"));
-        ChessGame.move(board, new Board.Position("e8"), new Board.Position("e5"));
-        Piece result2 = board.findPiece(new Board.Position("e5"));
-
-
-        ChessGame.move(board, new Board.Position("e5"), new Board.Position("a5"));
-        Piece result3 = board.findPiece(new Board.Position("a5"));
-        ChessGame.move(board, new Board.Position("a5"), new Board.Position("f5"));
-        Piece result4 = board.findPiece(new Board.Position("f5"));
-
-        assertThrows(
-                IllegalArgumentException.class,
-                () -> ChessGame.move(board, new Board.Position("f5"), new Board.Position("a1"))
-        );
-
-        // then
-        assertEquals(result1, Rook.createBlackRook());
-        assertEquals(result2, Rook.createBlackRook());
-        assertEquals(result3, Rook.createBlackRook());
-        assertEquals(result4, Rook.createBlackRook());
-        assertEquals(expect, BoardView.showBoard(board));
-    }
-
-
-    @Test
-    @DisplayName("나이트 기물의 움직임 테스트")
-    void moveKnight() {
-        // given
-        String knightSample =
-                "........" +
-                "........" +
-                "....N...." +
-                "........" +
-                "........" +
-                "........" +
-                "........" +
-                "........";
-
-        String expect =
-                "...N...." + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE;
-        ChessGame.initialize(board, knightSample);
-
-        // when
-        ChessGame.move(board, new Board.Position("e6"), new Board.Position("d8"));
-        Piece result1 = board.findPiece(new Board.Position("d8"));
-
-        assertThrows(
-                IllegalArgumentException.class,
-                () -> ChessGame.move(board, new Board.Position("d8"), new Board.Position("d7"))
-        );
-
-        // then
-        assertEquals(result1, Knight.createBlackKnight());
-        assertEquals(expect, BoardView.showBoard(board));
-    }
-
-    @Test
-    @DisplayName("비숍 기물의 움직임 테스트")
-    void moveBishop() {
-        // given
-        String bishopSample =
-                "........" +
-                "........" +
-                "....B..." +
-                "........" +
-                "........" +
-                "........" +
-                "........" +
-                "........";
-
-        String expect =
-                "......B." + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE;
-        ChessGame.initialize(board, bishopSample);
-
-        // when
-        ChessGame.move(board, new Board.Position("e6"), new Board.Position("g8"));
-        Piece result1 = board.findPiece(new Board.Position("g8"));
-
-        assertThrows(
-                IllegalArgumentException.class,
-                () -> ChessGame.move(board, new Board.Position("g8"), new Board.Position("g7"))
-        );
-
-        // then
-        assertEquals(result1, Bishop.createBlackBishop());
-        assertEquals(expect, BoardView.showBoard(board));
-    }
-
-    @Test
-    @DisplayName("없는 기물의 움직임 테스트는 실패해야 합니다")
-    void moveEmpty() {
-        // given
-        String expect =
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE;
-        ChessGame.initializeEmpty(board);
-
-        // when
-        assertThrows(
-                IllegalArgumentException.class,
-                () -> ChessGame.move(board, new Board.Position("e6"), new Board.Position("g7"))
-        );
-
-        // then
-        assertEquals(expect, BoardView.showBoard(board));
-    }
-
-    @Test
-    @DisplayName("폰 움직임 테스트")
-    void movePawn() {
-        // given
-        String sample =
-                "........" +
-                "........" +
-                "........" +
-                "........" +
-                "...Pp..." +
-                "........" +
-                "........" +
-                "........" ;
-
-        String expect =
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "....p..." + NEWLINE +
-                "........" + NEWLINE +
-                "...P...." + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE;
-        ChessGame.initialize(board, sample);
-
-        // when
-        assertThrows(
-                IllegalArgumentException.class,
-                () -> ChessGame.move(board, new Board.Position("d4"), new Board.Position("d5"))
-        );
-        assertThrows(
-                IllegalArgumentException.class,
-                () -> ChessGame.move(board, new Board.Position("d4"), new Board.Position("c4"))
-        );
-
-        assertThrows(
-                IllegalArgumentException.class,
-                () -> ChessGame.move(board, new Board.Position("d4"), new Board.Position("e4"))
-        );
-        ChessGame.move(board, new Board.Position("d4"), new Board.Position("d3"));
-        ChessGame.move(board, new Board.Position("e4"), new Board.Position("e5"));
-
-        // then
-        assertEquals(expect, BoardView.showBoard(board));
-    }
-
-    @Test
-    @DisplayName("앙파상 테스트")
-    void moveEnPassant() {
-        // given
-        String sample =
-                "........" +
-                "........" +
-                "........" +
-                "....p..." +
-                "...P...." +
-                "........" +
-                "........" +
-                "........" ;
-
-        String expect =
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "....P..." + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE +
-                "........" + NEWLINE;
-        ChessGame.initialize(board, sample);
-
-        // when
-        assertThrows(
-                IllegalArgumentException.class,
-                () -> ChessGame.move(board, new Board.Position("d4"), new Board.Position("c5"))
-        );
-        ChessGame.move(board, new Board.Position("d4"), new Board.Position("e5"));
-
-        // then
-        assertEquals(expect, BoardView.showBoard(board));
-    }
-
-
 }
