+++ src/test/java/softeer2nd/chess/pieces/KingTest.java
@@ -1,0 +1,4 @@
+package softeer2nd.chess.pieces;
+
+public class KingTest {
+}
